refactor(remediationorchestrator): Phase 6 Wave 2a - decompose Reconcile/NewReconciler, split oversized files
Extracts NewReconciler's inline AnalyzingCallbacks/AwaitingApprovalCallbacks
closures (which inflated its cyclomatic score) into named (r *Reconciler)
methods, and decomposes Reconcile's pending-phase bootstrap and terminal-phase
housekeeping into shouldSkipPendingReconcile, initializeNewRemediationRequest,
and handleTerminalPhaseHousekeeping (Reconcile: -> 14 cyclomatic).

Splits reconciler.go (1,148 lines) into reconciler.go (struct/NewReconciler/
timeout+routing construction), reconciler_callbacks.go (event/lock/RAR/hash
callbacks + config accessors), and reconcile_loop.go (Reconcile/pending-
bootstrap/terminal-housekeeping/SetupWithManager/field indexes).

Splits terminal_transitions.go (1,032 lines) into terminal_transitions.go
(inherited-completion/failure transitions, transitionPhase, verifying/failed
transitions), blocked_transitions.go (handleBlocked + escalation/notification/
metrics helpers), and timeout_handling.go (handleGlobalTimeout +
createEffectivenessAssessmentIfNeeded). Removes an orphaned doc comment left
over from an earlier refactor that had already moved populateTimeoutDefaults
to timeout_management.go.

All files now under the 700-line convention threshold. Per
GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520, see §7j-3).
