+++ internal/controller/remediationorchestrator/notification_creation.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -60,83 +61,101 @@ func (r *Reconciler) ensureNotificationsCreated(ctx context.Context, rr *remedia
 	logger := log.FromContext(ctx).WithValues("remediationRequest", rr.Name)
 
 	completionName := fmt.Sprintf("nr-completion-%s", rr.Name)
-	bulkName := fmt.Sprintf("nr-bulk-%s", rr.Name)
-
 	if !hasNotificationRef(rr, completionName) {
-		aiName := fmt.Sprintf("ai-%s", rr.Name)
-		ai := &aianalysisv1.AIAnalysis{}
-		if err := r.client.Get(ctx, client.ObjectKey{Name: aiName, Namespace: rr.Namespace}, ai); err != nil {
-			logger.Error(err, "Failed to fetch AIAnalysis for completion notification, will retry", "aiAnalysis", aiName)
+		r.createCompletionNotification(ctx, rr, completionName, logger)
+	}
+
+	bulkName := fmt.Sprintf("nr-bulk-%s", rr.Name)
+	if rr.Status.DuplicateCount > 0 && !hasNotificationRef(rr, bulkName) {
+		r.createBulkDuplicateNotification(ctx, rr, logger)
+	}
+}
+
+// createCompletionNotification gathers the AIAnalysis, WFE execution engine,
+// and EA verification summary (all best-effort/graceful-degradation) needed
+// to render the completion notification, then creates it (BR-ORCH-045).
+// Extracted from ensureNotificationsCreated per
+// GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520).
+func (r *Reconciler) createCompletionNotification(ctx context.Context, rr *remediationv1.RemediationRequest, completionName string, logger logr.Logger) {
+	aiName := fmt.Sprintf("ai-%s", rr.Name)
+	ai := &aianalysisv1.AIAnalysis{}
+	if err := r.client.Get(ctx, client.ObjectKey{Name: aiName, Namespace: rr.Namespace}, ai); err != nil {
+		logger.Error(err, "Failed to fetch AIAnalysis for completion notification, will retry", "aiAnalysis", aiName)
+		return
+	}
+
+	// Issue #518: Read executionEngine from WFE status (resolved at runtime by WE controller).
+	executionEngine := ""
+	if rr.Status.WorkflowExecutionRef != nil {
+		we := &workflowexecutionv1.WorkflowExecution{}
+		weKey := client.ObjectKey{Name: rr.Status.WorkflowExecutionRef.Name, Namespace: rr.Status.WorkflowExecutionRef.Namespace}
+		if weErr := r.client.Get(ctx, weKey, we); weErr != nil {
+			logger.V(1).Info("Could not fetch WFE for executionEngine (best-effort)", "error", weErr)
 		} else {
-			// Issue #518: Read executionEngine from WFE status (resolved at runtime by WE controller).
-			executionEngine := ""
-			if rr.Status.WorkflowExecutionRef != nil {
-				we := &workflowexecutionv1.WorkflowExecution{}
-				weKey := client.ObjectKey{Name: rr.Status.WorkflowExecutionRef.Name, Namespace: rr.Status.WorkflowExecutionRef.Namespace}
-				if weErr := r.client.Get(ctx, weKey, we); weErr != nil {
-					logger.V(1).Info("Could not fetch WFE for executionEngine (best-effort)", "error", weErr)
-				} else {
-					executionEngine = we.Status.ExecutionEngine
-				}
-			}
-			// #318: Fetch EA for verification summary (graceful degradation: nil if unavailable)
-			var ea *eav1.EffectivenessAssessment
-			if rr.Status.EffectivenessAssessmentRef != nil {
-				eaObj := &eav1.EffectivenessAssessment{}
-				eaKey := client.ObjectKey{
-					Name:      rr.Status.EffectivenessAssessmentRef.Name,
-					Namespace: rr.Namespace,
-				}
-				if err := r.client.Get(ctx, eaKey, eaObj); err != nil {
-					if apierrors.IsNotFound(err) {
-						logger.Info("EA not found for verification summary, notification will show 'not available'",
-							"ea", eaKey.Name)
-					} else {
-						logger.Error(err, "Failed to fetch EA for verification summary, notification will show 'not available'",
-							"ea", eaKey.Name)
-					}
-				} else {
-					ea = eaObj
-				}
-			}
-			notifName, notifErr := r.notificationCreator.CreateCompletionNotification(ctx, rr, ai, executionEngine, ea)
-			if notifErr != nil {
-				logger.Error(notifErr, "Failed to create completion notification, will retry")
-			} else {
-				logger.Info("Created completion notification", "notification", notifName)
-				ref := r.buildNotificationRef(ctx, notifName, rr.Namespace)
-				if refErr := helpers.UpdateRemediationRequestStatus(ctx, r.client, rr, func(rr *remediationv1.RemediationRequest) error {
-					rr.Status.NotificationRequestRefs = append(rr.Status.NotificationRequestRefs, ref)
-					return nil
-				}); refErr != nil {
-					logger.Error(refErr, "Failed to persist completion NT ref (non-critical)", "notification", notifName)
-				}
-				if r.Recorder != nil {
-					r.Recorder.Event(rr, corev1.EventTypeNormal, events.EventReasonNotificationCreated,
-						fmt.Sprintf("Completion notification created: %s", notifName))
-				}
-			}
+			executionEngine = we.Status.ExecutionEngine
 		}
 	}
 
-	if rr.Status.DuplicateCount > 0 && !hasNotificationRef(rr, bulkName) {
-		name, bulkErr := r.notificationCreator.CreateBulkDuplicateNotification(ctx, rr)
-		if bulkErr != nil {
-			logger.Error(bulkErr, "Failed to create bulk duplicate notification, will retry")
-		} else {
-			logger.Info("Created bulk duplicate notification", "notification", name)
-			ref := r.buildNotificationRef(ctx, name, rr.Namespace)
-			if refErr := helpers.UpdateRemediationRequestStatus(ctx, r.client, rr, func(rr *remediationv1.RemediationRequest) error {
-				rr.Status.NotificationRequestRefs = append(rr.Status.NotificationRequestRefs, ref)
-				return nil
-			}); refErr != nil {
-				logger.Error(refErr, "Failed to persist bulk NT ref (non-critical)", "notification", name)
-			}
-			if r.Recorder != nil {
-				r.Recorder.Event(rr, corev1.EventTypeNormal, events.EventReasonNotificationCreated,
-					fmt.Sprintf("Bulk duplicate notification created: %s", name))
-			}
+	// #318: Fetch EA for verification summary (graceful degradation: nil if unavailable)
+	var ea *eav1.EffectivenessAssessment
+	if rr.Status.EffectivenessAssessmentRef != nil {
+		eaObj := &eav1.EffectivenessAssessment{}
+		eaKey := client.ObjectKey{
+			Name:      rr.Status.EffectivenessAssessmentRef.Name,
+			Namespace: rr.Namespace,
 		}
+		if err := r.client.Get(ctx, eaKey, eaObj); err != nil {
+			if apierrors.IsNotFound(err) {
+				logger.Info("EA not found for verification summary, notification will show 'not available'",
+					"ea", eaKey.Name)
+			} else {
+				logger.Error(err, "Failed to fetch EA for verification summary, notification will show 'not available'",
+					"ea", eaKey.Name)
+			}
+		} else {
+			ea = eaObj
+		}
+	}
+
+	notifName, notifErr := r.notificationCreator.CreateCompletionNotification(ctx, rr, ai, executionEngine, ea)
+	if notifErr != nil {
+		logger.Error(notifErr, "Failed to create completion notification, will retry")
+		return
+	}
+	logger.Info("Created completion notification", "notification", notifName)
+	ref := r.buildNotificationRef(ctx, notifName, rr.Namespace)
+	if refErr := helpers.UpdateRemediationRequestStatus(ctx, r.client, rr, func(rr *remediationv1.RemediationRequest) error {
+		rr.Status.NotificationRequestRefs = append(rr.Status.NotificationRequestRefs, ref)
+		return nil
+	}); refErr != nil {
+		logger.Error(refErr, "Failed to persist completion NT ref (non-critical)", "notification", notifName)
+	}
+	if r.Recorder != nil {
+		r.Recorder.Event(rr, corev1.EventTypeNormal, events.EventReasonNotificationCreated,
+			fmt.Sprintf("Completion notification created: %s", notifName))
+	}
+}
+
+// createBulkDuplicateNotification creates the bulk-duplicate-signal
+// notification (BR-ORCH-034). Extracted from ensureNotificationsCreated per
+// GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520).
+func (r *Reconciler) createBulkDuplicateNotification(ctx context.Context, rr *remediationv1.RemediationRequest, logger logr.Logger) {
+	name, bulkErr := r.notificationCreator.CreateBulkDuplicateNotification(ctx, rr)
+	if bulkErr != nil {
+		logger.Error(bulkErr, "Failed to create bulk duplicate notification, will retry")
+		return
+	}
+	logger.Info("Created bulk duplicate notification", "notification", name)
+	ref := r.buildNotificationRef(ctx, name, rr.Namespace)
+	if refErr := helpers.UpdateRemediationRequestStatus(ctx, r.client, rr, func(rr *remediationv1.RemediationRequest) error {
+		rr.Status.NotificationRequestRefs = append(rr.Status.NotificationRequestRefs, ref)
+		return nil
+	}); refErr != nil {
+		logger.Error(refErr, "Failed to persist bulk NT ref (non-critical)", "notification", name)
+	}
+	if r.Recorder != nil {
+		r.Recorder.Event(rr, corev1.EventTypeNormal, events.EventReasonNotificationCreated,
+			fmt.Sprintf("Bulk duplicate notification created: %s", name))
 	}
 }
