+++ internal/controller/remediationorchestrator/verifying_handler.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
@@ -94,76 +95,16 @@ func (h *VerifyingHandler) Handle(ctx context.Context, rr *remediationv1.Remedia
 	}
 
 	if rr.Status.VerificationDeadline == nil {
-		eaName := rr.Status.EffectivenessAssessmentRef.Name
-		ea := &eav1.EffectivenessAssessment{}
-		if err := h.k8sClient.Get(ctx, client.ObjectKey{Name: eaName, Namespace: rr.Namespace}, ea); err != nil {
-			logger.Error(err, "Failed to fetch EA for VerificationDeadline computation", "ea", eaName)
-			return phase.Requeue(config.RequeueResourceBusy, "EA fetch error"), nil
-		}
-
-		if ea.Status.ValidityDeadline != nil {
-			deadline := metav1.NewTime(ea.Status.ValidityDeadline.Add(VerificationDeadlineBuffer))
-			if err := helpers.UpdateRemediationRequestStatus(ctx, h.k8sClient, rr, func(rr *remediationv1.RemediationRequest) error {
-				rr.Status.VerificationDeadline = &deadline
-				return nil
-			}); err != nil {
-				logger.Error(err, "Failed to persist VerificationDeadline")
-				return phase.TransitionIntent{}, err
-			}
-			logger.Info("VerificationDeadline set", "deadline", deadline.Format(time.RFC3339))
-		} else if time.Since(rr.CreationTimestamp.Time) > h.timeouts.Verifying {
-			logger.Info("Safety-net timeout: VerificationDeadline never set, RR exceeded verifying timeout",
-				"age", time.Since(rr.CreationTimestamp.Time).String(),
-				"verifyingTimeout", h.timeouts.Verifying.String())
-			if err := helpers.UpdateRemediationRequestStatus(ctx, h.k8sClient, rr, func(rr *remediationv1.RemediationRequest) error {
-				now := metav1.Now()
-				rr.Status.OverallPhase = phase.Completed
-				rr.Status.Outcome = "VerificationTimedOut"
-				rr.Status.CompletedAt = &now
-				rr.Status.ObservedGeneration = rr.Generation
-				return nil
-			}); err != nil {
-				logger.Error(err, "Failed to transition to Completed on safety-net timeout")
-				return phase.TransitionIntent{}, err
-			}
-			if h.m != nil {
-				h.m.PhaseTransitionsTotal.WithLabelValues(string(phase.Verifying), string(phase.Completed), rr.Namespace).Inc()
-			}
-			h.callbacks.EnsureNotificationsCreated(ctx, rr)
-			h.callbacks.EmitVerificationTimedOutAudit(ctx, rr)
-			if rr.Status.StartTime != nil {
-				h.callbacks.EmitCompletionAudit(ctx, rr, rr.Status.Outcome, time.Since(rr.Status.StartTime.Time).Milliseconds())
-			}
-			return phase.NoOp("safety-net timeout completed"), nil
-		} else {
-			logger.V(1).Info("EA.Status.ValidityDeadline not yet set, requeueing")
-			return phase.Requeue(config.RequeueResourceBusy, "ValidityDeadline not set"), nil
+		intent, done, err := h.ensureVerificationDeadline(ctx, rr, logger)
+		if done {
+			return intent, err
 		}
 	}
 
 	if rr.Status.VerificationDeadline != nil && time.Now().After(rr.Status.VerificationDeadline.Time) {
 		logger.Info("VerificationDeadline expired, timing out verification",
 			"deadline", rr.Status.VerificationDeadline.Format(time.RFC3339))
-		if err := helpers.UpdateRemediationRequestStatus(ctx, h.k8sClient, rr, func(rr *remediationv1.RemediationRequest) error {
-			now := metav1.Now()
-			rr.Status.OverallPhase = phase.Completed
-			rr.Status.Outcome = "VerificationTimedOut"
-			rr.Status.CompletedAt = &now
-			rr.Status.ObservedGeneration = rr.Generation
-			return nil
-		}); err != nil {
-			logger.Error(err, "Failed to transition to Completed on verification timeout")
-			return phase.TransitionIntent{}, err
-		}
-		if h.m != nil {
-			h.m.PhaseTransitionsTotal.WithLabelValues(string(phase.Verifying), string(phase.Completed), rr.Namespace).Inc()
-		}
-		h.callbacks.EnsureNotificationsCreated(ctx, rr)
-		h.callbacks.EmitVerificationTimedOutAudit(ctx, rr)
-		if rr.Status.StartTime != nil {
-			h.callbacks.EmitCompletionAudit(ctx, rr, rr.Status.Outcome, time.Since(rr.Status.StartTime.Time).Milliseconds())
-		}
-		return phase.NoOp("verification deadline expired, completed"), nil
+		return h.completeAsVerificationTimedOut(ctx, rr, logger, "verification deadline expired, completed")
 	}
 
 	if err := h.callbacks.TrackEffectivenessStatus(ctx, rr); err != nil {
@@ -180,4 +121,72 @@ func (h *VerifyingHandler) Handle(ctx context.Context, rr *remediationv1.Remedia
 	}
 
 	return phase.Requeue(config.RequeueResourceBusy, "EA still in progress"), nil
+}
+
+// ensureVerificationDeadline computes and persists VerificationDeadline from
+// the EA's ValidityDeadline (plus buffer) on first entry to Verifying with no
+// deadline set yet. If the EA hasn't computed ValidityDeadline and the RR has
+// exceeded the configured Verifying timeout, this applies a safety-net
+// completion (VerificationTimedOut) so the RR is never stuck indefinitely.
+// Extracted from Handle per GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520).
+// Returns done=true when the caller must return (intent, err) immediately.
+func (h *VerifyingHandler) ensureVerificationDeadline(ctx context.Context, rr *remediationv1.RemediationRequest, logger logr.Logger) (phase.TransitionIntent, bool, error) {
+	eaName := rr.Status.EffectivenessAssessmentRef.Name
+	ea := &eav1.EffectivenessAssessment{}
+	if err := h.k8sClient.Get(ctx, client.ObjectKey{Name: eaName, Namespace: rr.Namespace}, ea); err != nil {
+		logger.Error(err, "Failed to fetch EA for VerificationDeadline computation", "ea", eaName)
+		return phase.Requeue(config.RequeueResourceBusy, "EA fetch error"), true, nil
+	}
+
+	if ea.Status.ValidityDeadline != nil {
+		deadline := metav1.NewTime(ea.Status.ValidityDeadline.Add(VerificationDeadlineBuffer))
+		if err := helpers.UpdateRemediationRequestStatus(ctx, h.k8sClient, rr, func(rr *remediationv1.RemediationRequest) error {
+			rr.Status.VerificationDeadline = &deadline
+			return nil
+		}); err != nil {
+			logger.Error(err, "Failed to persist VerificationDeadline")
+			return phase.TransitionIntent{}, true, err
+		}
+		logger.Info("VerificationDeadline set", "deadline", deadline.Format(time.RFC3339))
+		return phase.TransitionIntent{}, false, nil
+	}
+
+	if time.Since(rr.CreationTimestamp.Time) > h.timeouts.Verifying {
+		logger.Info("Safety-net timeout: VerificationDeadline never set, RR exceeded verifying timeout",
+			"age", time.Since(rr.CreationTimestamp.Time).String(),
+			"verifyingTimeout", h.timeouts.Verifying.String())
+		intent, err := h.completeAsVerificationTimedOut(ctx, rr, logger, "safety-net timeout completed")
+		return intent, true, err
+	}
+
+	logger.V(1).Info("EA.Status.ValidityDeadline not yet set, requeueing")
+	return phase.Requeue(config.RequeueResourceBusy, "ValidityDeadline not set"), true, nil
+}
+
+// completeAsVerificationTimedOut transitions the RR to Completed with
+// outcome VerificationTimedOut, shared by the safety-net timeout (no
+// VerificationDeadline ever computed) and the normal deadline-expiry path.
+// Extracted from Handle per GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2
+// (issue #1520).
+func (h *VerifyingHandler) completeAsVerificationTimedOut(ctx context.Context, rr *remediationv1.RemediationRequest, logger logr.Logger, noOpReason string) (phase.TransitionIntent, error) {
+	if err := helpers.UpdateRemediationRequestStatus(ctx, h.k8sClient, rr, func(rr *remediationv1.RemediationRequest) error {
+		now := metav1.Now()
+		rr.Status.OverallPhase = phase.Completed
+		rr.Status.Outcome = "VerificationTimedOut"
+		rr.Status.CompletedAt = &now
+		rr.Status.ObservedGeneration = rr.Generation
+		return nil
+	}); err != nil {
+		logger.Error(err, "Failed to transition to Completed on verification timeout")
+		return phase.TransitionIntent{}, err
+	}
+	if h.m != nil {
+		h.m.PhaseTransitionsTotal.WithLabelValues(string(phase.Verifying), string(phase.Completed), rr.Namespace).Inc()
+	}
+	h.callbacks.EnsureNotificationsCreated(ctx, rr)
+	h.callbacks.EmitVerificationTimedOutAudit(ctx, rr)
+	if rr.Status.StartTime != nil {
+		h.callbacks.EmitCompletionAudit(ctx, rr, rr.Status.Outcome, time.Since(rr.Status.StartTime.Time).Milliseconds())
+	}
+	return phase.NoOp(noOpReason), nil
 }
